+++ packages/lib/test/treeUtils/sandbox.test.ts
@@ -35,6 +35,45 @@ test("sandbox creates instance of SandboxManager", () => {
   expect(manager instanceof SandboxManager).toBeTruthy()
 })
 
+test("withSandbox can be called with one node or a tuple of nodes", () => {
+  const a = new A({ b: new B({ value: 0 }) })
+  const manager = sandbox(a)
+  autoDispose(() => manager.dispose())
+
+  manager.withSandbox(a, node => {
+    assert(node, _ as A)
+    expect(node.$modelType).toBe("A")
+    return false
+  })
+
+  manager.withSandbox([a], nodes => {
+    assert(nodes, _ as [A])
+    expect(nodes[0].$modelType).toBe("A")
+    return false
+  })
+
+  manager.withSandbox([a, a], nodes => {
+    assert(nodes, _ as [A, A])
+    expect(nodes[0].$modelType).toBe("A")
+    expect(nodes[1].$modelType).toBe("A")
+    return false
+  })
+
+  manager.withSandbox([a, a.b], nodes => {
+    assert(nodes, _ as [A, B])
+    expect(nodes[0].$modelType).toBe("A")
+    expect(nodes[1].$modelType).toBe("B")
+    return false
+  })
+
+  manager.withSandbox([a.b, a], nodes => {
+    assert(nodes, _ as [B, A])
+    expect(nodes[0].$modelType).toBe("B")
+    expect(nodes[1].$modelType).toBe("A")
+    return false
+  })
+})
+
 test("withSandbox callback is called when node is a child of subtreeRoot", () => {
   const a = new A({ b: new B({ value: 0 }) })
   const manager = sandbox(a)